+++ cube/swiss/include/input.h
@@ -1,0 +1,8 @@
+#ifndef __INPUT_H
+#define __INPUT_H
+
+u16 padsButtonsHeld();
+s8 padsStickX();
+s8 padsStickY();
+
+#endif
