+++ cube/swiss/source/devices/fsp/deviceHandler-FSP.c
@@ -17,7 +17,6 @@
 #include "exi.h"
 #include "bba.h"
 #include "patcher.h"
-#include "util.h"
 
 extern int net_initialized;
 static FSP_SESSION *fsp_session;
