+++ cube/swiss/source/devices/ftp/deviceHandler-FTP.c
@@ -37,7 +37,6 @@
 #include "exi.h"
 #include "ftp_devoptab.h"
 #include "bba.h"
-#include "util.h"
 
 /* FTP Globals */
 extern int net_initialized;
